requirements.txt
et-xmlfile==1.1.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
tzdata==2023.3
